add interpret, compute_output_nifti and _compute_latent_tensor to predictor
