+++ clinicadl/utils/predict_manager/predict_manager.py
@@ -283,7 +283,7 @@ class PredictManager:
                     network=network,
                 )
             if save_nifti:
-                self.maps_manager._compute_output_nifti(
+                self._compute_output_nifti(
                     data_test,
                     data_group,
                     split,
@@ -292,7 +292,7 @@ class PredictManager:
                     network=network,
                 )
             if save_latent_tensor:
-                self.maps_manager._compute_latent_tensors(
+                self._compute_latent_tensors(
                     data_test,
                     data_group,
                     split,
@@ -369,7 +369,7 @@ class PredictManager:
                 gpu=gpu,
             )
         if save_nifti:
-            self.maps_manager._compute_output_nifti(
+            self._compute_output_nifti(
                 data_test,
                 data_group,
                 split,
@@ -377,10 +377,370 @@ class PredictManager:
                 gpu=gpu,
             )
         if save_latent_tensor:
-            self.maps_manager._compute_latent_tensors(
+            self._compute_latent_tensors(
                 data_test,
                 data_group,
                 split,
                 selection_metrics,
                 gpu=gpu,
             )
+
+    def _compute_latent_tensors(
+        self,
+        dataset,
+        data_group,
+        split,
+        selection_metrics,
+        nb_images=None,
+        gpu=None,
+        network=None,
+    ):
+        """
+        Compute the output tensors and saves them in the MAPS.
+
+        Args:
+            dataset (clinicadl.utils.caps_dataset.data.CapsDataset): wrapper of the data set.
+            data_group (str): name of the data group used for the task.
+            split (int): split number.
+            selection_metrics (list[str]): metrics used for model selection.
+            nb_images (int): number of full images to write. Default computes the outputs of the whole data set.
+            gpu (bool): If given, a new value for the device of the model will be computed.
+            network (int): Index of the network tested (only used in multi-network setting).
+        """
+        for selection_metric in selection_metrics:
+            # load the best trained model during the training
+            model, _ = self.maps_manager._init_model(
+                transfer_path=self.maps_manager.maps_path,
+                split=split,
+                transfer_selection=selection_metric,
+                gpu=gpu,
+                network=network,
+                nb_unfrozen_layer=self.maps_manager.nb_unfrozen_layer,
+            )
+            model = DDP(
+                model,
+                fsdp=self.maps_manager.fully_sharded_data_parallel,
+                amp=self.maps_manager.amp,
+            )
+            model.eval()
+
+            tensor_path = (
+                self.maps_manager.maps_path
+                / f"{self.maps_manager.split_name}-{split}"
+                / f"best-{selection_metric}"
+                / data_group
+                / "latent_tensors"
+            )
+            if cluster.master:
+                tensor_path.mkdir(parents=True, exist_ok=True)
+            dist.barrier()
+
+            if nb_images is None:  # Compute outputs for the whole data set
+                nb_modes = len(dataset)
+            else:
+                nb_modes = nb_images * dataset.elem_per_image
+
+            for i in [
+                *range(cluster.rank, nb_modes, cluster.world_size),
+                *range(int(nb_modes % cluster.world_size <= cluster.rank)),
+            ]:
+                data = dataset[i]
+                image = data["image"]
+                logger.debug(f"Image for latent representation {image}")
+                with autocast(enabled=self.maps_manager.std_amp):
+                    _, latent, _ = model.module._forward(
+                        image.unsqueeze(0).to(model.device)
+                    )
+                latent = latent.squeeze(0).cpu().float()
+                participant_id = data["participant_id"]
+                session_id = data["session_id"]
+                mode_id = data[f"{self.maps_manager.mode}_id"]
+                output_filename = f"{participant_id}_{session_id}_{self.maps_manager.mode}-{mode_id}_latent.pt"
+                torch.save(latent, tensor_path / output_filename)
+
+    @torch.no_grad()
+    def _compute_output_nifti(
+        self,
+        dataset,
+        data_group,
+        split,
+        selection_metrics,
+        gpu=None,
+        network=None,
+    ):
+        """
+        Computes the output nifti images and saves them in the MAPS.
+
+        Args:
+            dataset (clinicadl.utils.caps_dataset.data.CapsDataset): wrapper of the data set.
+            data_group (str): name of the data group used for the task.
+            split (int): split number.
+            selection_metrics (list[str]): metrics used for model selection.
+            gpu (bool): If given, a new value for the device of the model will be computed.
+            network (int): Index of the network tested (only used in multi-network setting).
+        # Raise an error if mode is not image
+        """
+        import nibabel as nib
+        from numpy import eye
+
+        for selection_metric in selection_metrics:
+            # load the best trained model during the training
+            model, _ = self.maps_manager._init_model(
+                transfer_path=self.maps_manager.maps_path,
+                split=split,
+                transfer_selection=selection_metric,
+                gpu=gpu,
+                network=network,
+                nb_unfrozen_layer=self.maps_manager.nb_unfrozen_layer,
+            )
+            model = DDP(
+                model,
+                fsdp=self.maps_manager.fully_sharded_data_parallel,
+                amp=self.maps_manager.amp,
+            )
+            model.eval()
+
+            nifti_path = (
+                self.maps_manager.maps_path
+                / f"{self.maps_manager.split_name}-{split}"
+                / f"best-{selection_metric}"
+                / data_group
+                / "nifti_images"
+            )
+            if cluster.master:
+                nifti_path.mkdir(parents=True, exist_ok=True)
+            dist.barrier()
+
+            nb_imgs = len(dataset)
+            for i in [
+                *range(cluster.rank, nb_imgs, cluster.world_size),
+                *range(int(nb_imgs % cluster.world_size <= cluster.rank)),
+            ]:
+                data = dataset[i]
+                image = data["image"]
+                x = image.unsqueeze(0).to(model.device)
+                with autocast(enabled=self.maps_manager.std_amp):
+                    output = model(x)
+                output = output.squeeze(0).detach().cpu().float()
+                # Convert tensor to nifti image with appropriate affine
+                input_nii = nib.Nifti1Image(image[0].detach().cpu().numpy(), eye(4))
+                output_nii = nib.Nifti1Image(output[0].numpy(), eye(4))
+                # Create file name according to participant and session id
+                participant_id = data["participant_id"]
+                session_id = data["session_id"]
+                input_filename = f"{participant_id}_{session_id}_image_input.nii.gz"
+                output_filename = f"{participant_id}_{session_id}_image_output.nii.gz"
+                nib.save(input_nii, nifti_path / input_filename)
+                nib.save(output_nii, nifti_path / output_filename)
+
+    def interpret(
+        self,
+        data_group,
+        name,
+        method,
+        caps_directory: Path = None,
+        tsv_path: Path = None,
+        split_list=None,
+        selection_metrics=None,
+        multi_cohort=False,
+        diagnoses=(),
+        target_node=0,
+        save_individual=False,
+        batch_size=None,
+        n_proc=None,
+        gpu=None,
+        amp=False,
+        overwrite=False,
+        overwrite_name=False,
+        level=None,
+        save_nifti=False,
+    ):
+        """
+        Performs the interpretation task on a subset of caps_directory defined in a TSV file.
+        The mean interpretation is always saved, to save the individual interpretations set save_individual to True.
+
+        Parameters
+        ----------
+        data_group: str
+            Name of the data group interpreted.
+        name: str
+            Name of the interpretation procedure.
+        method: str
+            Method used for extraction (ex: gradients, grad-cam...).
+        caps_directory: str (Path)
+            Path to the CAPS folder. For more information please refer to
+            [clinica documentation](https://aramislab.paris.inria.fr/clinica/docs/public/latest/CAPS/Introduction/).
+            Default will load the value of an existing data group.
+        tsv_path: str (Path)
+            Path to a TSV file containing the list of participants and sessions to test.
+            Default will load the DataFrame of an existing data group.
+        split_list: list of int
+            List of splits to interpret. Default perform interpretation on all splits available.
+        selection_metrics: list of str
+            List of selection metrics to interpret.
+            Default performs the interpretation on all selection metrics available.
+        multi_cohort: bool
+            If True considers that tsv_path is the path to a multi-cohort TSV.
+        diagnoses: list of str
+            List of diagnoses to load if tsv_path is a split_directory.
+            Default uses the same as in training step.
+        target_node: int
+            Node from which the interpretation is computed.
+        save_individual: bool
+            If True saves the individual map of each participant / session couple.
+        batch_size: int
+            If given, sets the value of batch_size, else use the same as in training step.
+        n_proc: int
+            If given, sets the value of num_workers, else use the same as in training step.
+        gpu: bool
+            If given, a new value for the device of the model will be computed.
+        amp: bool
+            If enabled, uses Automatic Mixed Precision (requires GPU usage).
+        overwrite: bool
+            If True erase the occurrences of data_group.
+        overwrite_name: bool
+            If True erase the occurrences of name.
+        level: int
+            Layer number in the convolutional part after which the feature map is chosen.
+        save_nifi : bool
+            If True, save the interpretation map in nifti format.
+        """
+
+        from clinicadl.interpret.gradients import method_dict
+
+        if method not in method_dict.keys():
+            raise NotImplementedError(
+                f"Interpretation method {method} is not implemented. "
+                f"Please choose in {method_dict.keys()}"
+            )
+
+        if not split_list:
+            split_list = self.maps_manager._find_splits()
+        logger.debug(f"List of splits {split_list}")
+
+        if self.maps_manager.multi_network:
+            raise NotImplementedError(
+                "The interpretation of multi-network framework is not implemented."
+            )
+
+        _, all_transforms = get_transforms(
+            normalize=self.maps_manager.normalize,
+            data_augmentation=self.maps_manager.data_augmentation,
+            size_reduction=self.maps_manager.size_reduction,
+            size_reduction_factor=self.maps_manager.size_reduction_factor,
+        )
+
+        group_df = None
+        if tsv_path is not None:
+            group_df = load_data_test(
+                tsv_path,
+                diagnoses if len(diagnoses) != 0 else self.maps_manager.diagnoses,
+                multi_cohort=multi_cohort,
+            )
+        self.maps_manager._check_data_group(
+            data_group, caps_directory, group_df, multi_cohort, overwrite
+        )
+
+        for split in split_list:
+            logger.info(f"Interpretation of split {split}")
+            df_group, parameters_group = self.maps_manager.get_group_info(
+                data_group, split
+            )
+
+            data_test = return_dataset(
+                parameters_group["caps_directory"],
+                df_group,
+                self.maps_manager.preprocessing_dict,
+                all_transformations=all_transforms,
+                multi_cohort=parameters_group["multi_cohort"],
+                label_presence=False,
+                label_code=self.maps_manager.label_code,
+                label=self.maps_manager.label,
+            )
+
+            test_loader = DataLoader(
+                data_test,
+                batch_size=batch_size
+                if batch_size is not None
+                else self.maps_manager.batch_size,
+                shuffle=False,
+                num_workers=n_proc if n_proc is not None else self.maps_manager.n_proc,
+            )
+
+            if not selection_metrics:
+                selection_metrics = self.maps_manager._find_selection_metrics(split)
+
+            for selection_metric in selection_metrics:
+                logger.info(f"Interpretation of metric {selection_metric}")
+                results_path = (
+                    self.maps_manager.maps_path
+                    / f"{self.maps_manager.split_name}-{split}"
+                    / f"best-{selection_metric}"
+                    / data_group
+                    / f"interpret-{name}"
+                )
+
+                if (results_path).is_dir():
+                    if overwrite_name:
+                        shutil.rmtree(results_path)
+                    else:
+                        raise MAPSError(
+                            f"Interpretation name {name} is already written. "
+                            f"Please choose another name or set overwrite_name to True."
+                        )
+                results_path.mkdir(parents=True)
+
+                model, _ = self.maps_manager._init_model(
+                    transfer_path=self.maps_manager.maps_path,
+                    split=split,
+                    transfer_selection=selection_metric,
+                    gpu=gpu,
+                )
+
+                interpreter = method_dict[method](model)
+
+                cum_maps = [0] * data_test.elem_per_image
+                for data in test_loader:
+                    images = data["image"].to(model.device)
+
+                    map_pt = interpreter.generate_gradients(
+                        images, target_node, level=level, amp=amp
+                    )
+                    for i in range(len(data["participant_id"])):
+                        mode_id = data[f"{self.maps_manager.mode}_id"][i]
+                        cum_maps[mode_id] += map_pt[i]
+                        if save_individual:
+                            single_path = (
+                                results_path
+                                / f"{data['participant_id'][i]}_{data['session_id'][i]}_{self.maps_manager.mode}-{data[f'{self.maps_manager.mode}_id'][i]}_map.pt"
+                            )
+                            torch.save(map_pt[i], single_path)
+                            if save_nifti:
+                                import nibabel as nib
+                                from numpy import eye
+
+                                single_nifti_path = (
+                                    results_path
+                                    / f"{data['participant_id'][i]}_{data['session_id'][i]}_{self.maps_manager.mode}-{data[f'{self.maps_manager.mode}_id'][i]}_map.nii.gz"
+                                )
+
+                                output_nii = nib.Nifti1Image(map_pt[i].numpy(), eye(4))
+                                nib.save(output_nii, single_nifti_path)
+
+                for i, mode_map in enumerate(cum_maps):
+                    mode_map /= len(data_test)
+
+                    torch.save(
+                        mode_map,
+                        results_path / f"mean_{self.maps_manager.mode}-{i}_map.pt",
+                    )
+                    if save_nifti:
+                        import nibabel as nib
+                        from numpy import eye
+
+                        output_nii = nib.Nifti1Image(mode_map.numpy(), eye(4))
+                        nib.save(
+                            output_nii,
+                            results_path
+                            / f"mean_{self.maps_manager.mode}-{i}_map.nii.gz",
+                        )
